+++ src/Cool/Exporter/VideoExportProcess.cpp
@@ -56,7 +56,7 @@ auto VideoExportProcess::estimated_remaining_time() -> float
     const auto nb_frames_to_render = static_cast<float>(_total_nb_of_frames_in_sequence - _nb_frames_sent_to_thread_pool);
 
     return nb_frames_to_render * _average_time_between_two_renders
-           + (_thread_pool.nb_jobs_in_queue() + _thread_pool.size() / 2.f) * _average_export_time / _thread_pool.size()
+           + (static_cast<float>(_thread_pool.nb_jobs_in_queue()) + static_cast<float>(_thread_pool.size()) / 2.f) * _average_export_time / static_cast<float>(_thread_pool.size())
            + 1.f;
 }
